requirements.txt
torch==2.7.0
torchrl==0.8.1
tensordict>=0.8.3,<0.9.0
hydra-core==1.3.2
wandb==0.19.11
tensorboard==2.19.0
grpcio==1.64.1
pygame==2.6.1
imageio==2.37.0
pytest==8.3.5
numpy==2.2.6